Kyle

+++ src/main/java/com/thefitnation/service/UserDemographicService.java
@@ -1,19 +1,28 @@
 package com.thefitnation.service;
 
+import com.thefitnation.domain.User;
 import com.thefitnation.domain.UserDemographic;
 import com.thefitnation.repository.UserDemographicRepository;
+import com.thefitnation.repository.UserRepository;
+import com.thefitnation.security.SecurityUtils;
 import com.thefitnation.service.dto.UserDemographicDTO;
 import com.thefitnation.service.mapper.UserDemographicMapper;
+import com.thefitnation.tools.AccountAuthTool;
+import com.thefitnation.web.rest.util.HeaderUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.stereotype.Service;
 
-import java.util.LinkedList;
-import java.util.List;
-import java.util.stream.Collectors;
+import java.util.ArrayList;
+import java.util.Objects;
+import java.util.Optional;
+
+import static org.hibernate.id.IdentifierGenerator.ENTITY_NAME;
 
 /**
  * Service Implementation for managing UserDemographic.
@@ -24,11 +33,14 @@ public class UserDemographicService {
 
     private final Logger log = LoggerFactory.getLogger(UserDemographicService.class);
 
+    private final UserRepository userRepository;
+
     private final UserDemographicRepository userDemographicRepository;
 
     private final UserDemographicMapper userDemographicMapper;
 
-    public UserDemographicService(UserDemographicRepository userDemographicRepository, UserDemographicMapper userDemographicMapper) {
+    public UserDemographicService(UserRepository userRepository, UserDemographicRepository userDemographicRepository, UserDemographicMapper userDemographicMapper) {
+        this.userRepository = userRepository;
         this.userDemographicRepository = userDemographicRepository;
         this.userDemographicMapper = userDemographicMapper;
     }
@@ -41,54 +53,96 @@ public class UserDemographicService {
      */
     public UserDemographicDTO save(UserDemographicDTO userDemographicDTO) {
         log.debug("Request to save UserDemographic : {}", userDemographicDTO);
+        User user = AccountAuthTool.getLoggedInUser(userRepository);
+        if (user == null) {
+            return null;
+        }
+        if (userDemographicDTO.getUserId() == null) {
+            userDemographicDTO.setUserId(user.getId());
+        }
         UserDemographic userDemographic = userDemographicMapper.userDemographicDTOToUserDemographic(userDemographicDTO);
-        userDemographic = userDemographicRepository.save(userDemographic);
+
+        if (AccountAuthTool.isAdmin(user)) {
+            UserDemographic dbUserDemographic = userDemographicRepository.findOneByUserWithEagerRelationships(userDemographic.getUser().getId());
+            if (dbUserDemographic == null) {
+                userDemographic.setId(null);
+            }
+            userDemographic = userDemographicRepository.save(userDemographic);
+        } else {
+            userDemographic.setUser(user);
+            UserDemographic dbUserDemographic = userDemographicRepository.findOneByUserWithEagerRelationships(user.getId());
+            if (dbUserDemographic == null) {
+                userDemographic.setId(null);
+            } else {
+                userDemographic.setId(dbUserDemographic.getId());
+            }
+            userDemographic = userDemographicRepository.save(userDemographic);
+        }
+
         UserDemographicDTO result = userDemographicMapper.userDemographicToUserDemographicDTO(userDemographic);
         return result;
     }
 
     /**
-     *  Get all the userDemographics.
+     * Get all the userDemographics.
      *
-     *  @param pageable the pagination information
-     *  @return the list of entities
+     * @param pageable the pagination information
+     * @return the list of entities
      */
     @Transactional(readOnly = true)
     public Page<UserDemographicDTO> findAll(Pageable pageable) {
         log.debug("Request to get all UserDemographics");
-        Page<UserDemographic> result = userDemographicRepository.findAll(pageable);
-        return result.map(userDemographic -> userDemographicMapper.userDemographicToUserDemographicDTO(userDemographic));
+        User user = AccountAuthTool.getLoggedInUser(userRepository);
+        Page<UserDemographic> result;
+        if (AccountAuthTool.isAdmin(user)) {
+            result = userDemographicRepository.findAll(pageable);
+        } else {
+            return null;
+        }
+        return result.map(userDemographicMapper::userDemographicToUserDemographicDTO);
     }
 
     /**
-     *  Get one userDemographic by id.
+     * Get one userDemographic by id.
      *
-     *  @param id the id of the entity
-     *  @return the entity
+     * @param id the id of the entity
+     * @return the entity
      */
     @Transactional(readOnly = true)
     public UserDemographicDTO findOne(Long id) {
         log.debug("Request to get UserDemographic : {}", id);
-        UserDemographic userDemographic = userDemographicRepository.findOneWithEagerRelationships(id);
-        UserDemographicDTO userDemographicDTO = userDemographicMapper.userDemographicToUserDemographicDTO(userDemographic);
-        return userDemographicDTO;
+        User user = AccountAuthTool.getLoggedInUser(userRepository);
+        UserDemographic userDemographic;
+        if (AccountAuthTool.isAdmin(user)) {
+            userDemographic = userDemographicRepository.findOneWithEagerRelationships(id);
+        } else {
+            userDemographic = userDemographicRepository.findOneWithEagerRelationships(id);
+            if (!Objects.equals(userDemographic.getUser().getId(), user.getId())) {
+                return null;
+            }
+        }
+        return userDemographicMapper.userDemographicToUserDemographicDTO(userDemographic);
     }
 
     @Transactional(readOnly = true)
-    public UserDemographicDTO findOneByUser(Long id) {
-        log.debug("Request to get UserDemographic by User : {}", id);
-        UserDemographic userDemographic = userDemographicRepository.findOneByUserWithEagerRelationships(id);
-        UserDemographicDTO userDemographicDTO = userDemographicMapper.userDemographicToUserDemographicDTO(userDemographic);
-        return userDemographicDTO;
+    public UserDemographicDTO findOneByUser() {
+        log.debug("Request to get UserDemographic : {}");
+        User user = AccountAuthTool.getLoggedInUser(userRepository);
+        UserDemographic userDemographic;
+        userDemographic = userDemographicRepository.findOneByUserWithEagerRelationships(user.getId());
+        return userDemographicMapper.userDemographicToUserDemographicDTO(userDemographic);
     }
 
     /**
-     *  Delete the  userDemographic by id.
+     * Delete the  userDemographic by id.
      *
-     *  @param id the id of the entity
+     * @param id the id of the entity
      */
     public void delete(Long id) {
         log.debug("Request to delete UserDemographic : {}", id);
-        userDemographicRepository.delete(id);
+        User user = AccountAuthTool.getLoggedInUser(userRepository);
+        if (AccountAuthTool.isAdmin(user)) {
+            userDemographicRepository.delete(id);
+        }
     }
 }
